Add todos Add default value in PlotSvgExport and get rid of one of methods.

+++ plot-base-portable/src/commonMain/kotlin/jetbrains/datalore/plot/base/stat/BoxplotStat.kt
@@ -30,7 +30,7 @@ import kotlin.math.sqrt
  */
 class BoxplotStat : BaseStat(DEF_MAPPING) {
 
-    private var myWhiskerIQRRatio: Double = 0.toDouble()          // ggplot: 'coef'
+    private var myWhiskerIQRRatio: Double = 0.0          // ggplot: 'coef'
     private var myComputeWidth = false    // ggplot: 'varWidth'
 
     /**
